+++ appdaemon/plugins/hass/exceptions.py
@@ -1,4 +1,9 @@
 
+from dataclasses import dataclass, field
+
+from appdaemon import exceptions as ade
+
+
 class HAAuthenticationError(Exception):
     pass
 
@@ -9,3 +14,17 @@ class HAEventsSubError(Exception):
 
 class HAFailedAuthentication(Exception):
     pass
+
+
+@dataclass
+class ScriptNotFound(ade.AppDaemonException):
+    script_name: str
+    namespace: str
+    plugin_name: str
+    domain: str = field(init=False, default="script")
+
+    def __str__(self):
+        res = f"'{self.script_name}' not found in plugin '{self.plugin_name}'"
+        if self.namespace != "default":
+            res += f" with namespace '{self.namespace}'"
+        return res
